refactor: shared 페이지 정적 생성으로 수정
